Add hub and light discovery, automatic pairing, and manual configuration to Philips Hue interface
Removes dependence on manually updating a JSON file to discover lights.
Allows users of this interface to press the hue hub pairing button to
automatically pair their hub during startup. Currently does not persist
the username retrieved from this process, forcing the user to enter the
config page to manually persist it.